Rename audioUtil to util, remove dependency eslint rule, fix dependency array in initializegame
