+++ .eslintrc.js
@@ -38,6 +38,7 @@ module.exports = {
     'react/jsx-no-bind': 'off',
     'jsx-a11y/media-has-caption': 'off',
     'import/prefer-default-export': 'off',
+    'react-hooks/exhaustive-deps': 'off',
   },
   parserOptions: {
     ecmaVersion: 2022,
